% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/straight-line.R
\name{full_on_audit}
\alias{full_on_audit}
\title{Walk a straight line}
\usage{
full_on_audit(quiet = TRUE)
}
\arguments{
\item{quiet}{if \code{TRUE} show no output during audit}
}
\value{
eventually a standardized audit object but for now a lynis audit
object which is a list with warnings, suggestions and full
audit results (data frame)
}
\description{
NOTE: This can take several minutes to run
}
\note{
Seriously, this can take several minutes to run
}
