% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign-commits.R
\name{gh_store_key}
\alias{gh_store_key}
\title{Add a public key to your GitHub account}
\usage{
gh_store_key(key, .token = NULL)
}
\arguments{
\item{key}{A character string containing the ID of a key to use. See
[gpg::gpg_list_keys()]; if you haven't created a key, see
[gpg::gpg_keygen()].}

\item{.token}{GitHub Personal Access Token.}
}
\description{
`gh_store_key()` add the public key associated with a key ID to your GitHub
account. If you have a GitHub Personal Access Token it will attempt to use
it; if it fails, it will print the public key for you to copy manually into
GitHub.
}
\details{
See
https://help.github.com/articles/creating-a-personal-access-token-for-the-command-line/
for more information on tokens.
}
\examples{
\dontrun{
newkey <- sign_commits_with_key("John Doe", "johndoe@example.com")
gh_store_key(newkey)
}
}
