% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-lynis.R
\name{print.lynis_obj}
\alias{print.lynis_obj}
\title{Print lynis audit results}
\usage{
\method{print}{lynis_obj}(x, ...)
}
\arguments{
\item{x}{lynis_obj}

\item{...}{unused}
}
\description{
Print lynis audit results
}
\keyword{internal}
