% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropsec-package.R, R/sobriety-package.R
\docType{package}
\name{ropsec}
\alias{ropsec}
\alias{ropsec-package}
\title{Personal Workstation Safety Check}
\description{
ropsec checkpoints are designed to help ensure personal and public
safety. Methods are provided to run "ropsec" checks on your system to ensure
your computing environment is as safe as possible from the perspectives of
confidentiality and integrity.

ropsec checkpoints are designed to help ensure personal and public
safety. Methods are provided to run "ropsec" checks on your system to ensure
your computing environment is as safe as possible from the perspectives of
confidentiality and integrity.
}
\author{
Ildiko Czeller (czeildi@gmail.com)

Karthik Ram (karthik.ram@gmail.com)

Bob Rudis (bob@rud.is)

Kara Woo

Bob Rudis (bob@rud.is)
}
