% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign-commits.R
\name{sign_commits_with_key}
\alias{sign_commits_with_key}
\title{GPG sign all git commits}
\usage{
sign_commits_with_key(name, email, passphrase = NULL, key = NULL)
}
\arguments{
\item{name}{A character string containing your name. If not provided,
`sign_commits_with_key()` will look in your global git configuration.}

\item{email}{A character string containing your email address. If not
provided, `sign_commits_with_key()` will look in your global git
configuration.}

\item{passphrase}{An optional passphrase to protect the keypair.}

\item{key}{A character string containing the ID of a pre-existing key to use.
If `NULL`, a new key will be created.}
}
\value{
A character string containing the ID of the key that was provided or
  generated.
}
\description{
Configure git to sign all commits using GPG. If no existing key is provided,
`sign_commits_with_key()` will create a new key and use it for signing.
}
\examples{
\dontrun{
newkey <- sign_commits_with_key("John Doe", "johndoe@example.com")
}
}
