% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_results.R
\name{summarize_system_checks}
\alias{summarize_system_checks}
\title{testthat style system checks}
\usage{
summarize_system_checks()
}
\value{
testthat report
}
\description{
Test system settings like ssh key, password requirement etc.
}
\examples{
\dontrun(summarize_system_check())
}
