% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpg-store.R
\name{gh_store_key}
\alias{gh_store_key}
\title{Add a public key to your GitHub account}
\usage{
gh_store_key(key, .token = NULL, open_url = is_interactive())
}
\arguments{
\item{key}{A character string containing the ID of a key to use, use the
return value of \code{sign_commits_with_key()}. If you are not sure, what key
you want to use, you can list your locally available keys with
\code{\link[gpg:gpg_list_keys]{gpg::gpg_list_keys()}}. The email address corresponding to this key must be
identical to the email address that you use with GitHub.}

\item{.token}{GitHub Personal Access Token with at least \code{write:gpg_key}
scope enabled. You can grant access to tokens
\href{https://github.com/settings/tokens}{here}.}

\item{open_url}{boolean, whether open relevant URLs automatically.}
}
\description{
\code{gh_store_key()} adds the public key associated with a key ID to your GitHub
account. If you have a GitHub Personal Access Token it will attempt to use
it; if it fails, it will print the public key for you to copy manually into
GitHub.
}
\details{
If you do not have a GitHub Personal Access Token setup or you want to store
your key on Gitlab or other service you can either call this function without
a token and then add the printed public key manually or call
\code{cat(}\code{\link[gpg:gpg_export]{gpg::gpg_export()}}\code{)} with \code{newkey} and add the returned public key
manually. See \href{https://help.github.com/articles/creating-a-personal-access-token-for-the-command-line/}{this page}
for more information on tokens.
}
\examples{
\dontrun{
new_key <- sign_commits_with_key("John Doe", "johndoe@example.com")
# if you do not have personal access token for github
gh_store_key(key = new_key)
# if your GitHub Personal Access Token is stored in `.Renviron` as GITHUB_PAT
gh_store_key(key = new_key, .token = Sys.getenv('GITHUB_PAT'))
}
}
